# More efficient parsing of BigInt values