Before:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                           (size)   Mode  Cnt         Score         Error  Units
[info] BigIntBenchmark.readJsoniterScala       10  thrpt    5  28609840.548 ± 1689063.362  ops/s
[info] BigIntBenchmark.readJsoniterScala      100  thrpt    5   1413079.020 ±    8176.779  ops/s
[info] BigIntBenchmark.readJsoniterScala     1000  thrpt    5     57806.661 ±    2977.655  ops/s
[info] BigIntBenchmark.readJsoniterScala    10000  thrpt    5       751.482 ±       1.143  ops/s
[info] BigIntBenchmark.readJsoniterScala   100000  thrpt    5         7.515 ±       0.645  ops/s
[info] BigIntBenchmark.readJsoniterScala  1000000  thrpt    5         0.075 ±       0.009  ops/s
```
After:
```
[info] REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
[info] why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
[info] experiments, perform baseline and negative tests that provide experimental control, make sure
[info] the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
[info] Do not assume the numbers tell you what you want them to tell.
[info] Benchmark                           (size)   Mode  Cnt         Score        Error  Units
[info] BigIntBenchmark.readJsoniterScala       10  thrpt    5  29684957.083 ± 918541.294  ops/s
[info] BigIntBenchmark.readJsoniterScala      100  thrpt    5   3258604.858 ± 281143.988  ops/s
[info] BigIntBenchmark.readJsoniterScala     1000  thrpt    5    121549.540 ±    946.224  ops/s
[info] BigIntBenchmark.readJsoniterScala    10000  thrpt    5      1420.101 ±      5.044  ops/s
[info] BigIntBenchmark.readJsoniterScala   100000  thrpt    5        14.582 ±      0.030  ops/s
[info] BigIntBenchmark.readJsoniterScala  1000000  thrpt    5         0.144 ±      0.004  ops/s
```